requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
astroid==3.0.1
black==23.11.0
click==8.1.7
crispy-bootstrap5==2023.10
dill==0.3.7
Django==4.2.7
django-apscheduler==0.6.2
django-crispy-forms==2.1
django-filter==23.4
django-stubs==4.2.6
django-stubs-ext==4.2.5
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
isort==5.12.0
mccabe==0.7.0
mypy==1.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pylint==3.0.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.3
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
tzlocal==5.2
